Finish creating ES import jest template
This is a template for a project that uses both Jest tests, and ES
modules and imports. Without this boilerplate, Jest will complain about
not knowing the word "import", or the symbol "{".

By cloning this repository, and filling out the package.json, you will
have a project that is ready for testing immediately.

Stack:
	- ECMAScript
	- Babel (pre-configured)
	- yarn